app/requirements.txt
Django==3.0.7
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework==3.12.4
ConfigParser
django-import-export
tqdm
djangorestframework-jwt==1.11.0
requests
pycryptodome
xmltodict
django-pandas==0.6.4
# 필요하지만 안되는것들.....
#numpy==1.16.0
#pandas==0.23.4 # 이놈 왜 설치안됨??... ㅡㅡ
#scikit-learn #sklearn은 안됨